Change Duke to Gina
Chatbot is the same as the given example.

A more unique chatbot personality would be more engaging for users.

Let's change
* the dialogue from the chatbot
* the display pictures for the chatbot and user

Doing so will represent Gina's personality better.
