--- src/main/java/duke/Launcher.java
@@ -1,9 +1,0 @@
-package duke;
-
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Gui.class, args);
-    }
-}
