--- src/main/java/duke/commands/DeleteCommand.java
@@ -1,44 +1,0 @@
-package duke.commands;
-
-import duke.DukeException;
-import duke.Storage;
-import duke.task.Task;
-import duke.TaskList;
-import duke.Ui;
-
-/**
- * Represents a command to delete a task.
- */
-public class DeleteCommand extends Command {
-    private final String input;
-
-    /**
-     * Constructs a command to delete a specified task.
-     *
-     * @param input The user input for the index of a task.
-     */
-    public DeleteCommand(String input) {
-        this.input = input;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public String execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        try {
-            int index = Integer.parseInt(input.trim()) - 1;
-            Task deletedTask = taskList.deleteTask(index);
-            storage.save(taskList);
-            return ui.showDeleteTask(deletedTask, taskList);
-        } catch (NumberFormatException e) {
-            throw new DukeException("Input a valid number!");
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public boolean isExit() {
-        return false;
-    }
-}
