--- src/main/java/duke/commands/ExitCommand.java
@@ -1,25 +1,0 @@
-package duke.commands;
-
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
-
-/**
- * Represents a command to exit the program.
- */
-public class ExitCommand extends Command {
-    /**
-     * {@inheritDoc}
-     */
-    public String execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        return ui.showExit();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public boolean isExit() {
-        return true;
-    }
-}
