--- src/main/java/duke/commands/ListCommand.java
@@ -1,25 +1,0 @@
-package duke.commands;
-
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
-
-/**
- * Represents a command to list current tasks.
- */
-public class ListCommand extends Command {
-    /**
-     * {@inheritDoc}
-     */
-    public String execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        return ui.showTaskList(taskList);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public boolean isExit() {
-        return false;
-    }
-}
