--- src/main/java/duke/commands/MarkCommand.java
@@ -1,43 +1,0 @@
-package duke.commands;
-
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
-
-/**
- * Represents a command to mark a task as done.
- */
-public class MarkCommand extends Command {
-    private String input;
-
-    /**
-     * Constructs a command to mark a specified task as done.
-     *
-     * @param input The user input for task number.
-     */
-    public MarkCommand(String input) {
-        this.input = input;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public String execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        try {
-            int index = Integer.parseInt(input.trim()) - 1;
-            taskList.get(index).markAsDone();
-            storage.save(taskList);
-            return ui.showMarkedTask(taskList.get(index));
-        } catch (NumberFormatException e) {
-            throw new DukeException("Input a valid number!");
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public boolean isExit() {
-        return false;
-    }
-}
