+++ src/main/java/gina/Launcher.java
@@ -1,0 +1,9 @@
+package gina;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Gui.class, args);
+    }
+}
